reafactor: refactore di setup
