+++ Src/Peresentation/Api/Config/Services/DependencyInjection.cs
@@ -1,0 +1,22 @@
+﻿using Api.Config.Services.Setup;
+using Application;
+
+namespace Api.Config.Services;
+
+public static class DependencyInjection
+{
+    public static WebApplicationBuilder AddServices(this WebApplicationBuilder builder)
+    {
+        builder.Services
+
+            .AddApplicationHandlers()
+
+            .AddPersistenceSetup(builder.Configuration)
+
+            .AddEndpointsApiExplorer()
+
+            .AddGraphQlSetup();
+
+        return builder;
+    }
+}
