+++ Src/Peresentation/Api/Config/Services/Setup/PersistenceSetup.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Persistence;
+
+namespace Api.Config.Services.Setup;
+
+public static class PersistenceSetup
+{
+    public static IServiceCollection AddPersistenceSetup(this IServiceCollection services, IConfiguration config)
+    {
+        var connectionString = config.GetConnectionString("SqlServer");
+
+        void Action(DbContextOptionsBuilder o) => o.UseSqlServer(connectionString);
+
+        services.AddReadRepositories(Action);
+        services.AddWriteRepositories(Action);
+
+        return services;
+    }
+}
