+++ Src/Peresentation/Api/Program.cs
@@ -1,4 +1,5 @@
 using Api.Config;
+using Api.Config.Services;
 
 WebApplication
     .CreateBuilder(args)
